# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/genetics": {
      "get": {
        "description": "Returns the full genetics reference catalog ordered alphabetically by Hebrew name. The list is intended to be cached client-side.",
        "operationId": "GeneticsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Catalog fetched successfully. `result` is an array of GeneticsDto ordered alphabetically by name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneticsListResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not applicable for this endpoint."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all genetics rows",
        "summaryHe": "מציגים את קטלוג הגנטיקה והזנים המלא במערכת",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-tree-evergreen"
      },
      "post": {
        "description": "Creates a new genetics entry in the reference catalog. The `name` field must be unique. Returns the created record wrapped in ServiceResultContainer.",
        "operationId": "GeneticsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Genetics creation payload. `name` and `color` are required. `color` must be a valid hex color (e.g., #228B22). `type` must be one of: היברידי, סאטיבה, אינדיקה.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneticsCreateDto"
              },
              "examples": {
                "basic": {
                  "summary": "Basic creation",
                  "value": {
                    "name": "גורילה גלו",
                    "description": "זן חזק במיוחד...",
                    "parent1": "Chem Sis",
                    "parent2": "Sour Dubb",
                    "origin": "ארה\"ב",
                    "type": "היברידי",
                    "color": "#228B22"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Genetics created successfully. `result` contains the created record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneticsResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (e.g., invalid hex color, missing required fields)."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "409": {
            "description": "A genetics with this name already exists."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new genetics record",
        "summaryHe": "יוצרים גנטיקה חדשה בקטלוג המערכת",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-tree-evergreen"
      }
    },
    "/genetics/{name}": {
      "get": {
        "description": "Returns the genetics row whose `name` column exactly matches the supplied `:name` path parameter. Result is `null` when no record matches — this is not treated as a 404 so the frontend can render a graceful empty state.",
        "operationId": "GeneticsController_findOne",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew strain name. Must exactly match the `genetics.name` column. URL-encode spaces and non-ASCII characters (e.g. `גורילה%20גלו`).",
            "schema": {
              "example": "גורילה גלו",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lookup succeeded. `result` is the matching strain or null when no row matches the given name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneticsResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not applicable for this endpoint."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "Not applicable for this endpoint. Missing rows return `result: null`."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a strain by name",
        "summaryHe": "מציגים פרטים מלאים על זן גנטיקה ספציפי לפי שמו",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-tree-evergreen"
      },
      "patch": {
        "description": "Updates an existing genetics entry in the reference catalog. All fields are optional — only provided fields are modified. The `name` path parameter identifies the record to update and cannot be changed.",
        "operationId": "GeneticsController_update",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew strain name. Must exactly match the `genetics.name` column. URL-encode spaces and non-ASCII characters.",
            "schema": {
              "example": "גורילה גלו",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Genetics update payload. All fields are optional. `color` must be a valid hex color if provided. `type` must be one of: היברידי, סאטיבה, אינדיקה. Pass `null` or empty string to clear optional string fields; pass empty array `[]` to clear effects/tags.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneticsUpdateDto"
              },
              "examples": {
                "partial": {
                  "summary": "Partial update",
                  "value": {
                    "description": "Updated description",
                    "color": "#1A5C1A"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Genetics updated successfully. `result` contains the updated record.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneticsResultResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (e.g., invalid hex color)."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "No genetics matches the given name."
          },
          "500": {
            "description": "Unexpected database or server error."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a genetics record by name",
        "summaryHe": "מעדכנים את מאפייני הגנטיקה של זן קיים לפי שמו",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-tree-evergreen"
      },
      "delete": {
        "operationId": "GeneticsController_delete",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew strain name to delete.",
            "schema": {
              "example": "גורילה גלו",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Strain deleted successfully."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "404": {
            "description": "No genetics matches the given name."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a genetics strain by name.",
        "summaryHe": "מחוקים זן גנטיקה לצמיות מקטלוג",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-trash"
      }
    },
    "/genetics/{name}/enrich": {
      "post": {
        "description": "Searches the web for the given strain name, sends context to LLM, and returns enriched description with web results. Does not persist — caller decides whether to save.",
        "operationId": "GeneticsController_enrichSingle",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Hebrew strain name to enrich.",
            "schema": {
              "example": "גורילה גלו",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrichment succeeded. `result` contains the LLM-generated description and web search results."
          },
          "400": {
            "description": "Invalid strain name."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "403": {
            "description": "Not applicable for this endpoint."
          },
          "404": {
            "description": "No genetics matches the given name."
          },
          "500": {
            "description": "LLM or web search failure."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enrich a single genetics record",
        "summaryHe": "מעשירים זן גנטיקה בודד בפרטים ונתוני מעבדה מבוססי AI",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-tree-evergreen"
      }
    },
    "/genetics/enrich-missing": {
      "post": {
        "operationId": "GeneticsController_enrichMissing",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bulk enrichment completed."
          },
          "401": {
            "description": "Missing or expired JWT token."
          },
          "500": {
            "description": "LLM or web search failure."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enrich all genetics with missing properties (thcRange, terpenes, effects).",
        "summaryHe": "מפעילים סריקה והעשרה אוטומטית לכל הזנים שחסר להם מידע מבוסס AI",
        "tags": [
          "genetics"
        ],
        "toolIcon": "ph-magic-wand"
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "GeneticsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Auto-incremented primary key.",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Hebrew display name. Unique — used as the lookup key from the frontend.",
            "example": "גורילה גלו"
          },
          "description": {
            "type": "string",
            "description": "Short Hebrew description of the strain. Omitted when not set.",
            "example": "זן חזק במיוחד שזכה במקומות ראשונים ב-Cannabis Cup..."
          },
          "parent1": {
            "type": "string",
            "description": "First genetic parent (parent1) — Hebrew or English name. Omitted when not set.",
            "example": "Chem Sis"
          },
          "parent2": {
            "type": "string",
            "description": "Second genetic parent (parent2) — Hebrew or English name. Omitted when not set.",
            "example": "Sour Dubb"
          },
          "origin": {
            "type": "string",
            "description": "Country / region of origin (Hebrew text). Omitted when not set.",
            "example": "ארה\"ב"
          },
          "type": {
            "type": "string",
            "description": "Cannabis type — היברידי / סאטיבה / אינדיקה. Omitted when not set.",
            "enum": [
              "היברידי",
              "סאטיבה",
              "אינדיקה"
            ],
            "example": "היברידי"
          },
          "thcRange": {
            "type": "string",
            "description": "THC percentage range (e.g. \"15-21%\"). Omitted when not set.",
            "example": "15-21%"
          },
          "terpenes": {
            "type": "string",
            "description": "Comma-separated list of dominant terpenes. Omitted when not set.",
            "example": "Caryophyllene, Limonene, Myrcene"
          },
          "effects": {
            "type": "string",
            "description": "Comma-separated list of effect labels in Hebrew. Omitted when not set.",
            "example": "מרגיעה, מרדימה, משככת כאבים"
          },
          "color": {
            "type": "string",
            "description": "Hex accent color used by the frontend to render per-strain UI accents (dots, tag borders, etc.).",
            "example": "#228B22"
          },
          "colorDark": {
            "type": "string",
            "description": "WCAG AA-safe variant of `color` for dark theme backgrounds (#080D1A).",
            "example": "#228B22"
          },
          "colorLight": {
            "type": "string",
            "description": "WCAG AA-safe variant of `color` for light theme backgrounds (#F0F4F8).",
            "example": "#1B6B1B"
          }
        },
        "required": [
          "id",
          "name",
          "color",
          "colorDark",
          "colorLight"
        ]
      },
      "GeneticsListResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable status message.",
            "example": "OK"
          },
          "result": {
            "description": "The list of genetics rows ordered alphabetically by name.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeneticsDto"
            }
          },
          "error": {
            "description": "Optional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "retryAfter": {
            "type": "string",
            "description": "Optional retry hint."
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code echoed from the server.",
            "example": 200
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "GeneticsResultResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the request succeeded.",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable status message.",
            "example": "OK"
          },
          "result": {
            "description": "The matching genetics row, or null when no record matches the given name.",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GeneticsDto"
              }
            ]
          },
          "error": {
            "description": "Optional error details.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "retryAfter": {
            "type": "string",
            "description": "Optional retry hint."
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code echoed from the server.",
            "example": 200
          }
        },
        "required": [
          "success",
          "message",
          "result"
        ]
      },
      "GeneticsCreateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Hebrew display name. Must be unique. Used as the lookup key from the frontend.",
            "example": "גורילה גלו"
          },
          "description": {
            "type": "string",
            "description": "Short Hebrew description of the strain.",
            "example": "זן חזק במיוחד שזכה במקומות ראשונים ב-Cannabis Cup..."
          },
          "parent1": {
            "type": "string",
            "description": "First genetic parent (parent1) — Hebrew or English name.",
            "example": "Chem Sis"
          },
          "parent2": {
            "type": "string",
            "description": "Second genetic parent (parent2) — Hebrew or English name.",
            "example": "Sour Dubb"
          },
          "origin": {
            "type": "string",
            "description": "Country / region of origin (Hebrew text).",
            "example": "ארה\"ב"
          },
          "type": {
            "type": "string",
            "description": "Cannabis type. Must be one of: היברידי, סאטיבה, אינדיקה.",
            "enum": [
              "היברידי",
              "סאטיבה",
              "אינדיקה"
            ],
            "example": "היברידי"
          },
          "color": {
            "type": "string",
            "description": "Hex accent color used by the frontend to render per-strain UI accents (dots, tag borders, etc.). Must be a valid hex color starting with #.",
            "example": "#228B22"
          }
        },
        "required": [
          "name",
          "color"
        ]
      },
      "GeneticsUpdateDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Short Hebrew description of the strain. Pass null or empty string to clear.",
            "example": "זן חזק במיוחד שזכה במקומות ראשונים ב-Cannabis Cup..."
          },
          "parent1": {
            "type": "string",
            "description": "First genetic parent (parent1) — Hebrew or English name. Pass null or empty string to clear.",
            "example": "Chem Sis"
          },
          "parent2": {
            "type": "string",
            "description": "Second genetic parent (parent2) — Hebrew or English name. Pass null or empty string to clear.",
            "example": "Sour Dubb"
          },
          "origin": {
            "type": "string",
            "description": "Country / region of origin (Hebrew text). Pass null or empty string to clear.",
            "example": "ארה\"ב"
          },
          "type": {
            "type": "string",
            "description": "Cannabis type. Must be one of: היברידי, סאטיבה, אינדיקה. Pass null to clear.",
            "enum": [
              "היברידי",
              "סאטיבה",
              "אינדיקה"
            ],
            "example": "היברידי"
          },
          "thcRange": {
            "type": "string",
            "description": "THC percentage range (e.g. \"15-21%\").",
            "example": "15-21%"
          },
          "terpenes": {
            "type": "string",
            "description": "Comma-separated list of dominant terpenes.",
            "example": "Caryophyllene, Limonene, Myrcene"
          },
          "effects": {
            "type": "string",
            "description": "Comma-separated list of effect labels in Hebrew.",
            "example": "מרגיעה, מרדימה, משככת כאבים"
          },
          "color": {
            "type": "string",
            "description": "Hex accent color used by the frontend to render per-strain UI accents (dots, tag borders, etc.). Must be a valid hex color starting with #.",
            "example": "#228B22"
          }
        }
      },
...
    }
  }
}